feat: add custom roles
