+++ packages/better-auth/src/plugins/access/index.ts
@@ -1,0 +1,2 @@
+export * from "./access";
+export * from "./types";
