+++ packages/better-auth/src/plugins/admin/access/statement.ts
@@ -1,4 +1,4 @@
-import { createAccessControl } from "./access";
+import { createAccessControl } from "../../access";
 
 export const defaultStatements = {
 	user: [
