+++ packages/better-auth/src/plugins/admin/client.ts
@@ -1,13 +1,8 @@
-import type { admin } from ".";
 import { BetterAuthError } from "../../error";
 import type { BetterAuthClientPlugin } from "../../types";
-import {
-	adminAc,
-	userAc,
-	type AccessControl,
-	type defaultStatements,
-	type Role,
-} from "./access";
+import { type AccessControl, type Role } from "../access";
+import { adminAc, defaultRoles, defaultStatements, userAc } from "./access";
+import type { admin } from "./admin";
 
 interface AdminClientOptions {
 	ac: AccessControl;
@@ -16,64 +11,72 @@ interface AdminClientOptions {
 	};
 }
 
-export const adminClient = <O extends AdminClientOptions>(options?: O) => {
-	type DefaultStatements = typeof defaultStatements;
-	type Statements = O["ac"] extends AccessControl<infer S>
-		? S extends Record<string, Array<any>>
-			? S & DefaultStatements
-			: DefaultStatements
-		: DefaultStatements;
-	const roles = {
-		admin: adminAc,
-		user: userAc,
-		...options?.roles,
-	};
+export const adminClient = Object.assign(
+	<O extends AdminClientOptions>(options?: O) => {
+		type DefaultStatements = typeof defaultStatements;
+		type Statements = O["ac"] extends AccessControl<infer S>
+			? S extends Record<string, Array<any>>
+				? S & DefaultStatements
+				: DefaultStatements
+			: DefaultStatements;
+		const roles = {
+			admin: adminAc,
+			user: userAc,
+			...options?.roles,
+		};
 
-	return {
-		id: "better-auth-client",
-		$InferServerPlugin: {} as ReturnType<
-			typeof admin<{
-				ac: O["ac"] extends AccessControl
-					? O["ac"]
-					: AccessControl<DefaultStatements>;
-				roles: O["roles"] extends Record<string, Role>
-					? O["roles"]
-					: {
-							admin: Role;
-							user: Role;
+		return {
+			id: "better-auth-client",
+			$InferServerPlugin: {} as ReturnType<
+				typeof admin<{
+					ac: O["ac"] extends AccessControl
+						? O["ac"]
+						: AccessControl<DefaultStatements>;
+					roles: O["roles"] extends Record<string, Role>
+						? O["roles"]
+						: {
+								admin: Role;
+								user: Role;
+							};
+				}>
+			>,
+			getActions: ($fetch) => ({
+				admin: {
+					checkRolePermission: <
+						R extends O extends { roles: any }
+							? keyof O["roles"]
+							: "admin" | "user",
+					>(data: {
+						role: R;
+						permission: {
+							//@ts-expect-error fix this later
+							[key in keyof Statements]?: Statements[key][number][];
 						};
-			}>
-		>,
-		getActions: ($fetch) => ({
-			admin: {
-				checkRolePermission: <
-				R extends O extends { roles: any }
-					? keyof O["roles"]
-					: "admin" | "user",
-			>(data: {
-				role: R;
-				permission: {
-					//@ts-expect-error fix this later
-					[key in keyof Statements]?: Statements[key][number][];
-				};
-			}) => {
-				if (Object.keys(data.permission).length > 1) {
-					throw new BetterAuthError(
-						"you can only check one resource permission at a time.",
-					);
-				}
-				const role = roles[data.role as unknown as "admin"];
-				if (!role) {
-					return false;
-				}
-				const isAuthorized = role?.authorize(data.permission as any);
-				return isAuthorized.success;
+					}) => {
+						if (Object.keys(data.permission).length > 1) {
+							throw new BetterAuthError(
+								"you can only check one resource permission at a time.",
+							);
+						}
+						const role = roles[data.role as unknown as "admin"];
+						if (!role) {
+							return false;
+						}
+						const isAuthorized = role?.authorize(data.permission as any);
+						return isAuthorized.success;
+					},
+				},
+			}),
+			pathMethods: {
+				"/admin/list-users": "GET",
+				"/admin/stop-impersonating": "POST",
 			},
-			}
-		}),
-		pathMethods: {
-			"/admin/list-users": "GET",
-			"/admin/stop-impersonating": "POST",
-		},
-	} satisfies BetterAuthClientPlugin;
-};
+		} satisfies BetterAuthClientPlugin;
+	},
+	{
+		defaultStatements,
+		defaultRoles,
+		adminAc,
+		userAc,
+	},
+);
