+++ packages/better-auth/src/plugins/organization/access/index.ts
@@ -1,3 +1,1 @@
-export * from "./access";
-export * from "./types";
-export * from "./statement";
+export * from './statement'
