+++ packages/better-auth/src/plugins/organization/organization.test.ts
@@ -3,7 +3,7 @@ import { getTestInstance } from "../../test-utils/test-instance";
 import { organization } from "./organization";
 import { createAuthClient } from "../../client";
 import { organizationClient } from "./client";
-import { createAccessControl } from "./access";
+import { createAccessControl } from "../access";
 import { ORGANIZATION_ERROR_CODES } from "./error-codes";
 import { BetterAuthError } from "../../error";
